fix firestore
