+++ data/src/main/java/com/whitebeach/data/remote/firestore/FireStoreRepository.kt
@@ -3,11 +3,10 @@ package com.whitebeach.data.remote.firestore
 import com.whitebeach.data.model.PlayerInfo
 
 interface FireStoreRepository {
-    suspend fun getPlayers(): List<PlayerInfo>
+    suspend fun getPlayers(): List<PlayerInfo?>
 }
 
 class FireStoreRepositoryImpl(): FireStoreRepository {
-    override suspend fun getPlayers(): List<PlayerInfo> =
+    override suspend fun getPlayers(): List<PlayerInfo?> =
         playerFireStoreDataSource()
-
 }
